requirements.txt
flask
python-dotenv
httpx
nest-asyncio
playwright
cf-clearance2